pagectl: split key
